refactor(bot): rename AppwriteService to DatabaseService; drop fallbacks
Replaces the 2120-line dual-backend AppwriteService with a 700-line
DatabaseService that only speaks Postgres + R2 + Redis. Every method
preserves its public signature so the 30+ consumers don't care about
the rename — the in-place bulk s/AppwriteService/DatabaseService/ in
the follow-up commits keeps them working.

Notable behavior changes:

- DATABASE_URL and the R2 credentials are now required at construction.
  Missing either throws a clear error instead of silently degrading to
  an Appwrite fallback that no longer exists.
- All the if (this.guildConfigRepo) { ... } else { Appwrite fallback }
  branches collapse to a single Postgres path. Less code, one shape to
  reason about.
- `realtimeClient` (Appwrite's websocket client) is gone. The
  subscribeToModules / subscribeToGuildConfigs methods now unconditionally
  subscribe via EventBus; no-op when Redis isn't configured.
- Legacy helpers that only made sense with Appwrite Storage —
  uploadRecordingFile(buffer, name), getRecordingFileUrl(fileId),
  getRecordingFileBuffer(nonR2Key) — are removed. The two remaining
  call sites in recording.ts already use the streaming R2 helpers.
- All method bodies delegate directly to @modus/db repositories; the
  inline TTL cache + EventBus publishes stay exactly where they were.
